Move filter query and RowComparer into UI base
